Extract shared dedupe/isObject utilities
- add dedupeStrings and isObject to packages/shared utils module
- replace duplicated local helpers in server notification and serve command modules
- replace duplicated dedupe helper in web push notifications hook
- export and test shared utils to keep behavior consistent across packages
